Added Cardslide
Cardslide is a random task, each time you swipe the card it has a 1 in 50 chance of being valid. The user of course wouldn't know this and thinks they need to be precise.
I also updated FixWiring and added wires but made it so wires will lock on with their matching wires so it's easier. Upload data is shorted by 10 seconds.